fixed and filled out many of the type declarations for all of the new types removed obsolete restart_ptl.[ch]

+++ src/adios_mpi.c
@@ -28,8 +28,8 @@ struct adios_MPI_data_struct
     int rank;
     int size;
     void * buffer;
-    unsigned long long buffer_size;
-    unsigned long long start;
+    uint64_t buffer_size;
+    uint64_t start;
     int last_var_write_yes; // was the last item asked to write a write="yes"?
 };
 
@@ -181,8 +181,8 @@ void adios_mpi_write (struct adios_file_struct * fd
 
     if (v->copy_on_write == adios_flag_yes)
     {
-        unsigned long long var_size;
-        unsigned long long mem_allowed;
+        uint64_t var_size;
+        uint64_t mem_allowed;
 
         var_size = adios_size_of_var (v, data);
         mem_allowed = adios_method_buffer_alloc (var_size);
@@ -212,12 +212,12 @@ void adios_mpi_write (struct adios_file_struct * fd
 
 void adios_mpi_get_write_buffer (struct adios_file_struct * fd
                                 ,struct adios_var_struct * v
-                                ,unsigned long long * size
+                                ,uint64_t * size
                                 ,void ** buffer
                                 ,struct adios_method_struct * method
                                 )
 {
-    unsigned long long mem_allowed;
+    uint64_t mem_allowed;
 
     if (*size == 0)
     {
@@ -285,9 +285,9 @@ static void adios_mpi_do_read (struct adios_file_struct * fd
                                                       method->method_data; 
     struct adios_var_struct * v = 0;
 
-    unsigned long long my_data_len = adios_data_size (fd->group);
+    uint64_t my_data_len = adios_data_size (fd->group);
     //printf("read: adios_data_size:%llu\n",my_data_len);
-    unsigned long long my_offset = 0;
+    uint64_t my_offset = 0;
     MPI_Status mstatus;
     int amode = MPI_MODE_RDONLY;
     int err;
@@ -356,7 +356,7 @@ static void adios_mpi_do_read (struct adios_file_struct * fd
                      );
             if (next != -1)
             {
-                unsigned long long new_offset = my_offset + my_data_len;
+                uint64_t new_offset = my_offset + my_data_len;
                 MPI_Isend (&new_offset, 1, MPI_LONG_LONG, next
                           ,current, group_comm, &md->req
                           );
@@ -381,8 +381,8 @@ static void adios_mpi_do_read (struct adios_file_struct * fd
 
     if (md->buffer_size < my_data_len)
     {
-        unsigned long long mem_needed = 0;
-        unsigned long long mem_allowed = 0;
+        uint64_t mem_needed = 0;
+        uint64_t mem_allowed = 0;
         mem_needed = my_data_len - md->buffer_size;
         mem_allowed = adios_method_buffer_alloc (mem_needed);
         if (mem_needed != mem_allowed)
@@ -421,13 +421,13 @@ static void adios_mpi_do_write (struct adios_file_struct * fd
     struct adios_MPI_data_struct * md = (struct adios_MPI_data_struct *)
                                                        method->method_data;
     char name [STR_LEN];
-    unsigned long long my_data_len = adios_data_size (fd->group);
-    unsigned long long mem_needed = 0;
-    unsigned long long mem_allowed = 0;
+    uint64_t my_data_len = adios_data_size (fd->group);
+    uint64_t mem_needed = 0;
+    uint64_t mem_allowed = 0;
     MPI_Offset my_offset = 0;
     MPI_Status mstatus;
     int overflow = 0;
-    unsigned long long end = 0;
+    uint64_t end = 0;
     struct adios_var_struct * v = fd->group->vars;
     struct adios_attribute_struct * a = fd->group->attributes;
     int err;
@@ -528,7 +528,7 @@ static void adios_mpi_do_write (struct adios_file_struct * fd
                      );
             if (next != -1)
             {
-                unsigned long long new_offset = my_offset + my_data_len;
+                uint64_t new_offset = my_offset + my_data_len;
                 MPI_Isend (&new_offset, 1, MPI_LONG_LONG, next
                           ,current, group_comm, &md->req
                           );
